+++ packages/gestalt/src/SelectList.tsx
@@ -50,9 +50,18 @@ type Props = {
    */
   name?: string;
   /**
+   * Callback triggered when the user blurs the input.
+   */
+  onBlur?: (arg1: { event: React.FocusEvent<HTMLSelectElement>; value: string }) => void;
+  /**
+  /**
    * Callback triggered when the user selects a new option.  See the [controlled component](https://gestalt.pinterest.systems/web/selectlist#Controlled-component) variant to learn more.
    */
   onChange: (arg1: { event: React.ChangeEvent<HTMLSelectElement>; value: string }) => void;
+  /**
+   * Callback triggered when the user focuses the input.
+   */
+  onFocus?: (arg1: { event: React.FocusEvent<HTMLSelectElement>; value: string }) => void;
   /**
    * If not provided, the first item in the list will be shown. Be sure to localize the text. See the [controlled component](https://gestalt.pinterest.systems/web/selectlist#Controlled-component) variant to learn more.
    */
@@ -84,7 +93,9 @@ function SelectList({
   label,
   labelDisplay = 'visible',
   name,
+  onBlur,
   onChange,
+  onFocus,
   placeholder,
   size = 'md',
   value,
@@ -97,11 +108,26 @@ function SelectList({
     }
   };
 
+  const handleBlur = (event: React.FocusEvent<HTMLSelectElement>) => {
+    const { value: eventValue } = event.target;
+    onBlur?.({ event, value: eventValue });
+    handleOnChange(event);
+    setFocused(false);
+  };
+
+  const handleFocus = (event: React.FocusEvent<HTMLSelectElement>) => {
+    const { value: eventValue } = event.target;
+    onFocus?.({ event, value: eventValue });
+    setFocused(true);
+  };
+
   const classes = classnames(
     styles.select,
     formElement.base,
     size === 'md' ? layout.medium : layout.large,
     {
+      [formElement.md]: size === 'md',
+      [formElement.lg]: size === 'lg',
       [formElement.normal]: !errorMessage,
       [formElement.enabledTransparent]: !disabled,
       [formElement.disabled]: disabled,
@@ -123,7 +149,7 @@ function SelectList({
 
   return (
     <Box>
-      {label && <FormLabel id={id} label={label} labelDisplay={labelDisplay} />}
+      {label && <FormLabel id={id} label={label} labelDisplay={labelDisplay} size={size} />}
       <Box
         dangerouslySetInlineStyle={{
           __style: { backgroundColor: TOKEN_COLOR_BACKGROUND_FORMFIELD_PRIMARY },
@@ -160,12 +186,9 @@ function SelectList({
           disabled={disabled}
           id={id}
           name={name}
-          onBlur={(event) => {
-            setFocused(false);
-            handleOnChange(event);
-          }}
+          onBlur={handleBlur}
           onChange={handleOnChange}
-          onFocus={() => setFocused(true)}
+          onFocus={handleFocus}
           // @ts-expect-error - TS2322 - Type 'string | null | undefined' is not assignable to type 'string | number | readonly string[] | undefined'.
           value={showPlaceholder ? placeholder : value}
         >